Overtime changes: Switch to just BTrees, since in the end, hash tables not only failed to be any faster, they were SLOWER. Probably due to more cache reads. In the version we handed in, we sort of ignored our experimental results by using hash tables, and that's bad.
Also changes a syntax error (for what appears to be dead code, but still
worth fixing probably) and made it delete answers file, and made it not
throw an exception when you try to delete a database that doesn't exist.

HEADS UP: Files that currently matter are mydbtest.py and DatabaseFunctions.py,
everything else is just junk.